Splitted question renderer to separated classes
