+++ src/components/editor/page-properties.tsx
@@ -1,3 +1,14 @@
+import { errorAtom } from "../../atoms/error";
+import { layoutsAtom } from "../../atoms/layouts";
+import { pagesAtom } from "../../atoms/pages";
+import { Page, PageProperty, PagePropertyType } from "../../generated/client";
+import { useApi } from "../../hooks/use-api";
+import strings from "../../localization/strings";
+import { LayoutType, QuestionType } from "../../types";
+import GenericDialog from "../generic/generic-dialog";
+import WithDebounce from "../generic/with-debounce";
+import { Close, Edit } from "@mui/icons-material";
+import AddCircleIcon from "@mui/icons-material/AddCircle";
 import {
   Box,
   Button,
@@ -7,19 +18,8 @@ import {
   TextField,
   Typography
 } from "@mui/material";
-import strings from "../../localization/strings";
-import { Close, Edit } from "@mui/icons-material";
-import { LayoutType, QuestionType } from "../../types";
-import AddCircleIcon from "@mui/icons-material/AddCircle";
-import { ChangeEvent, useEffect, useState } from "react";
 import { useAtom, useSetAtom } from "jotai";
-import GenericDialog from "../generic/generic-dialog";
-import { pagesAtom } from "../../atoms/pages";
-import { useApi } from "../../hooks/use-api";
-import { Page, PageProperty, PagePropertyType } from "../../generated/client";
-import { layoutsAtom } from "../../atoms/layouts";
-import WithDebounce from "../generic/with-debounce";
-import { errorAtom } from "../../atoms/error";
+import { ChangeEvent, useEffect, useState } from "react";
 
 /**
  * Component properties
@@ -131,18 +131,18 @@ const PageProperties = ({ pageNumber, surveyId }: Props) => {
       (property) => property.type === PagePropertyType.Options
     )
       ? surveyPages[pageNumber - 1].properties!.map((property) => {
-          return property.type === PagePropertyType.Options
-            ? { ...property, value: JSON.stringify(updatedOptions) }
-            : property;
-        })
+        return property.type === PagePropertyType.Options
+          ? { ...property, value: JSON.stringify(updatedOptions) }
+          : property;
+      })
       : [
-          ...surveyPages[pageNumber - 1].properties!,
-          {
-            key: PagePropertyType.Options,
-            value: JSON.stringify(updatedOptions),
-            type: PagePropertyType.Options
-          }
-        ];
+        ...surveyPages[pageNumber - 1].properties!,
+        {
+          key: PagePropertyType.Options,
+          value: JSON.stringify(updatedOptions),
+          type: PagePropertyType.Options
+        }
+      ];
 
     const updatesToPage: Page = {
       ...surveyPages[pageNumber - 1],
@@ -277,7 +277,7 @@ const PageProperties = ({ pageNumber, surveyId }: Props) => {
             <Typography variant="h6">
               {strings.editSurveysScreen.editPagesPanel.answerOptions}
             </Typography>
-            {options?.map((option, i) =>
+            {options.map((option, i) =>
               renderOptionsWithDebounceTextField(
                 "option",
                 (e) => handleQuestionOptionChange(e, option),
@@ -293,17 +293,17 @@ const PageProperties = ({ pageNumber, surveyId }: Props) => {
         (layout) =>
           layout.id === surveyPages[pageNumber - 1].layoutId && layout.name === LayoutType.QUESTION
       ) && (
-        <Box p={2} sx={{ borderBottom: "1px solid #DADCDE" }}>
-          <Button
-            size="large"
-            variant="text"
-            startIcon={<AddCircleIcon />}
-            onClick={addNewQuestionOption}
-          >
-            {strings.editSurveysScreen.editPagesPanel.addOption}
-          </Button>
-        </Box>
-      )}
+          <Box p={2} sx={{ borderBottom: "1px solid #DADCDE" }}>
+            <Button
+              size="large"
+              variant="text"
+              startIcon={<AddCircleIcon />}
+              onClick={addNewQuestionOption}
+            >
+              {strings.editSurveysScreen.editPagesPanel.addOption}
+            </Button>
+          </Box>
+        )}
     </>
   );
 };
