+++ src/utils/QuestionUtils.ts
@@ -1,0 +1,6 @@
+/**
+ * Abstract class for QuestionRenderer
+ */
+export abstract class AbstractQuestionRenderer {
+  public abstract render(options: string[]): string;
+}
